reorg: move generic code under Ema.Route.Generic
