+++ src/Ema/Example/Ex00_Hello.hs
@@ -5,8 +5,8 @@
 module Ema.Example.Ex00_Hello where
 
 import Ema
-import Ema.Multi.Generic
-import Ema.Multi.Generic.Motley
+import Ema.Route.Generic
+import Ema.Route.Generic.Sub
 import Generics.SOP qualified as SOP
 
 data Route = Route_Index
