+++ src/Ema/Example/Ex02_Clock.hs
@@ -16,9 +16,9 @@ import Data.List ((!!))
 import Data.Time (UTCTime, defaultTimeLocale, formatTime, getCurrentTime)
 import Ema
 import Ema.Example.Common (tailwindLayout)
-import Ema.Multi.Generic
-import Ema.Multi.Generic.Motley
 import Ema.Route.Encoder (RouteEncoder)
+import Ema.Route.Generic
+import Ema.Route.Generic.Sub
 import Generics.SOP qualified as SOP
 import Text.Blaze.Html5 ((!))
 import Text.Blaze.Html5 qualified as H
