+++ src/Ema/Example/Ex03_Store.hs
@@ -12,9 +12,9 @@ import Data.SOP (I (I), NP (Nil, (:*)))
 import Data.Text qualified as T
 import Ema
 import Ema.Example.Common (tailwindLayout, watchDirForked)
-import Ema.Multi.Generic
-import Ema.Multi.Generic.Motley
 import Ema.Route.Encoder
+import Ema.Route.Generic
+import Ema.Route.Generic.Sub
 import Generics.SOP qualified as SOP
 import Optics.Core (Iso', iso)
 import System.FSNotify qualified as FSNotify
